+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/catalog/customization/CalalogModifier.kt
@@ -1,0 +1,7 @@
+package com.miam.kmm_miam_sdk.android.ui.components.catalog.customization
+
+import androidx.compose.ui.Modifier
+
+object CatalogModifier {
+    var categoryListContainer = Modifier
+}